embed_error, critical bug fixes, code formatting
I fixed a couple of critical bugs in the code and will be pushing my new work directly to main to prevent double-work in fixing them.
Error 1: Logic error in check_emoji resulted in valid unicode emojis being rejected.
Error 2: embed_error sent the entire error which would result in a massive text block being sent when explaining to the user what went wrong.

I also re-formatted some of the longer blocks with Black to make them more readable. I didn't use Black on all of the code, as I'm personally not a fan of how strict it is, but it helped make some of the longer embed commands readable.

In this commit I've also pushed the updated embed_error which will provide the user with much more information about the error that was returned. I've accounted for the following errors:
MissingRequiredArgument
ArgumentParsingError
UnexpectedQuoteError
InvalidEndOfQuotedStringError
ExpectedClosingQuoteError
BadArgument
TooManyArguments
MissingPermissions/MissingRole/MissingAnyRole ---- Alerts the user that only admins can use /embed
Emoji usage errors as defined in check_emoji and embed_error_check

In comments, I've provided a brief explanation of each of these errors in their respective if-statement, along with the link to that error or that error's parent in the Discord.Py docs